# Improve async session updates 