--- a/cli/lint/schema.go
+++ b/cli/lint/schema.go
@@ -491,6 +491,9 @@ const confSchema = `{
 	"optimisations_use_async_session_write": {
 		"type": "boolean"
 	},
+	"session_update_pool_size":{
+		"type": "integer"
+	},
 	"pid_file_location": {
 		"type": "string"
 	},