@joshblakeley 2018-06-06
Currently panics on gateway shutdown so have that to fix still.