@buger 2018-06-06
Can you put some initial benchmarks to understand the difference it makes?