@buger 2018-06-07
Also if it will show good performance, and we decide to turn async session write by default, it will make sense remove old setting, and introduce new one: `disable_async_session_writes`